+++ turkin.nikolai/I2/repeatedNums.h
@@ -5,7 +5,7 @@ namespace turkin
 {
   struct RepeatedNums
   {
-    bool operator()(const int * array, size_t size, int begin, int end);
+    bool operator()(const int * array, size_t size, size_t begin, size_t end);
   };
 }
 #endif
